+++ src/main/java/com/kekie6/colorfulazaleas/decorators/ColorfulTreeDecorator.java
@@ -1,4 +1,4 @@
-package com.kekie6.colorfulazaleas.util;
+package com.kekie6.colorfulazaleas.decorators;
 
 import com.kekie6.colorfulazaleas.ColorfulAzaleas;
 import com.mojang.serialization.Codec;
@@ -6,9 +6,7 @@ import com.mojang.serialization.codecs.RecordCodecBuilder;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
-import net.minecraft.core.Registry;
 import net.minecraft.core.Vec3i;
-import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.levelgen.feature.stateproviders.*;
 import net.minecraft.world.level.levelgen.feature.treedecorators.TreeDecorator;
 import net.minecraft.world.level.levelgen.feature.treedecorators.TreeDecoratorType;
