+++ src/main/java/com/kekie6/colorfulazaleas/registry/ColorfulAzaleasItemGroups.java
@@ -1,7 +1,6 @@
-package com.kekie6.colorfulazaleas.util;
+package com.kekie6.colorfulazaleas.registry;
 
 import com.kekie6.colorfulazaleas.*;
-import com.kekie6.colorfulazaleas.registry.*;
 import net.fabricmc.fabric.api.itemgroup.v1.*;
 import net.minecraft.core.*;
 import net.minecraft.core.registries.*;
